refs #9 presentaion modularization
