+++ caclulator/starter/lib/domain/domain.dart
@@ -1,0 +1,3 @@
+export 'entity/entity.dart';
+export 'repository/repository.dart';
+export '../util/use_case.dart';
