+++ caclulator/starter/lib/domain/repository/repository.dart
@@ -1,0 +1,10 @@
+
+
+import 'package:calculator_modularization_presentation_starter/domain/domain.dart';
+
+abstract class ICalculatorRepository {
+
+  Future<CalculatorEntity> fetch();
+
+  Future<void> save(CalculatorEntity entity);
+}
