+++ caclulator/starter/lib/util/use_case.dart
@@ -1,0 +1,3 @@
+abstract class IUseCase<T, Params> {
+  Future<T> execute([Params params]);
+}
