+++ caclulator/starter/presentation/lib/presentation.dart
@@ -1,0 +1,3 @@
+library presentation;
+
+export 'src/presentation.dart';
