+++ caclulator/starter/presentation/lib/src/presentation.dart
@@ -1,0 +1,1 @@
+export 'view_model/calculator.dart';
