+++ caclulator/starter/presentation/pubspec.yaml
@@ -1,4 +1,4 @@
-name: calculator_modularization_presentation_starter
+name: calculator_modularization_presentation
 description: A new Flutter project.
 version: 1.0.0+1
 publish_to: none
@@ -10,12 +10,9 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
-  cupertino_icons: 1.0.6
-  shared_preferences: 2.2.2
-  provider: 6.1.1
 
-  calculator_modularization_presentation:
-    path: presentation
+  calculator_modularization_presentation_starter:
+    path: ../
 
 dev_dependencies:
   flutter_lints: 3.0.1
